자잘한 구조 수정 및 네이밍 리펙토링
