+++ .idea/encodings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/src/main/java/유성구.csv" charset="EUC-KR" />
+    <file url="file://$PROJECT_DIR$/src/main/java/유성구_외국인소비정보_2020.csv" charset="EUC-KR" />
+  </component>
+</project>
