+++ src/main/java/Control/DBUpdateControl.java
@@ -2,13 +2,14 @@ package Control;
 
 import Entity.Data;
 import com.opencsv.exceptions.CsvException;
+import persistence.InvalidNameFormatException;
 
 import java.io.File;
 import java.io.IOException;
 
 public class DBUpdateControl {
-    public void DBUpdateRequest(File csv) throws IOException, CsvException {
-        Data data = new Data();
-        data.DBUpdate(csv);
+    public void DBUpdateRequest(File csv) throws IOException, CsvException, InvalidNameFormatException {
+        Data data = new Data(csv);
+        data.DBUpdate();
     }
 }
