+++ src/main/java/persistence/DTO/ConsumptionAmountDTO.java
@@ -1,4 +1,4 @@
-package DB.DTO;
+package persistence.DTO;
 
 import com.opencsv.exceptions.CsvException;
 import lombok.Getter;
@@ -10,16 +10,17 @@ import java.io.IOException;
 @Getter
 @Setter
 @ToString
-public class DailyFloatingPopulationDTO extends DTO {
+public class ConsumptionAmountDTO extends DTO {
     private int id;
     private int year;
     private int month;
     private String dong_code;
     private String dong_name;
-    private String time;
-    private double daily_floating_population;
+    private String industry_code;
+    private String industry_name;
+    private double amount;
 
-    public DailyFloatingPopulationDTO(String[] line) throws IOException, CsvException {
+    public ConsumptionAmountDTO(String[] line) throws IOException, CsvException {
         super(line);
     }
 }
