+++ src/main/java/persistence/DTO/ConsumptionAmountForeignerDTO.java
@@ -1,4 +1,4 @@
-package DB.DTO;
+package persistence.DTO;
 
 import com.opencsv.exceptions.CsvException;
 import lombok.Getter;
@@ -15,16 +15,17 @@ public class ConsumptionAmountForeignerDTO extends DTO {
     private int year;
     private int month;
     private String metropolitan_code;
+    private String metropolitan_name;
     private String city_code;
+    private String city_name;
     private String dong_code;
     private String dong_name;
     private String industry_code;
     private String industry_name;
     private String nationality;
     private String period;
-    private int count;
+    private double count;
     private double amount;
-    private String consumption_amount_foreigner;
 
     public ConsumptionAmountForeignerDTO(String[] line) throws IOException, CsvException {
         super(line);
