+++ src/main/java/persistence/DTO/ConsumptionAmountOutsiderDTO.java
@@ -1,7 +1,6 @@
-package DB.DTO;
+package persistence.DTO;
 
 import com.opencsv.exceptions.CsvException;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -11,7 +10,7 @@ import java.io.IOException;
 @Getter
 @Setter
 @ToString
-public class ConsumptionAmountDTO extends DTO {
+public class ConsumptionAmountOutsiderDTO extends DTO {
     private int id;
     private int year;
     private int month;
@@ -19,9 +18,11 @@ public class ConsumptionAmountDTO extends DTO {
     private String dong_name;
     private String industry_code;
     private String industry_name;
+    private String residential_city_code;
+    private String residential_city_name;
     private double amount;
 
-    public ConsumptionAmountDTO(String[] line) throws IOException, CsvException {
+    public ConsumptionAmountOutsiderDTO(String[] line) throws IOException, CsvException {
         super(line);
     }
 }
