+++ src/main/java/persistence/DTO/DailyFloatingPopulationDTO.java
@@ -1,4 +1,4 @@
-package DB.DTO;
+package persistence.DTO;
 
 import com.opencsv.exceptions.CsvException;
 import lombok.Getter;
@@ -10,18 +10,16 @@ import java.io.IOException;
 @Getter
 @Setter
 @ToString
-public class ConsumptionAmountOutsiderDTO extends DTO {
+public class DailyFloatingPopulationDTO extends DTO {
     private int id;
     private int year;
     private int month;
     private String dong_code;
     private String dong_name;
-    private String industry_code;
-    private String industry_name;
-    private String residential_city;
-    private double amount;
+    private String time;
+    private double daily_floating_population;
 
-    public ConsumptionAmountOutsiderDTO(String[] line) throws IOException, CsvException {
+    public DailyFloatingPopulationDTO(String[] line) throws IOException, CsvException {
         super(line);
     }
 }
