+++ src/main/java/persistence/InvalidNameFormatException.java
@@ -1,0 +1,8 @@
+package persistence;
+
+public class InvalidNameFormatException extends Exception {
+    public InvalidNameFormatException() {}
+    public InvalidNameFormatException(String message) {
+        super(message);
+    }
+}
