+++ src/main/java/persistence/MyBatisConnectionFactory.java
@@ -1,4 +1,4 @@
-package DB;
+package persistence;
 
 import org.apache.ibatis.io.Resources;
 import org.apache.ibatis.session.SqlSessionFactory;
