+++ src/main/resources/config/config.xml
@@ -5,10 +5,10 @@
 
 <configuration>
     <typeAliases>
-        <typeAlias alias="CADTO" type="DB.DTO.ConsumptionAmountDTO"/>
-        <typeAlias alias="CAFDTO" type="DB.DTO.ConsumptionAmountForeignerDTO"/>
-        <typeAlias alias="CAODTO" type="DB.DTO.ConsumptionAmountOutsiderDTO"/>
-        <typeAlias alias="DFPDTO" type="DB.DTO.DailyFloatingPopulationDTO"/>
+        <typeAlias alias="CADTO" type="persistence.DTO.ConsumptionAmountDTO"/>
+        <typeAlias alias="CAFDTO" type="persistence.DTO.ConsumptionAmountForeignerDTO"/>
+        <typeAlias alias="CAODTO" type="persistence.DTO.ConsumptionAmountOutsiderDTO"/>
+        <typeAlias alias="DFPDTO" type="persistence.DTO.DailyFloatingPopulationDTO"/>
     </typeAliases>
 
     <environments default="development">
