+++ src/main/resources/sqlmapper/consumptionAmount.xml
@@ -4,7 +4,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="mapper.CAMapper">
-    <select id="selectAll" resultType="DB.DTO.ConsumptionAmountDTO">
+    <select id="selectAll" resultType="persistence.DTO.ConsumptionAmountDTO">
         SELECT *
         FROM consumption_amount
     </select>
