+++ src/main/resources/sqlmapper/consumptionAmountForeigner.xml
@@ -4,13 +4,13 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="mapper.CAFMapper">
-    <select id="selectAll" resultType="DB.DTO.ConsumptionAmountForeignerDTO">
+    <select id="selectAll" resultType="persistence.DTO.ConsumptionAmountForeignerDTO">
         SELECT *
         FROM consumption_amount_foreigner
     </select>
 
-    <insert id="insert" useGeneratedKeys="true" keyProperty="id" parameterType="DB.DTO.ConsumptionAmountForeignerDTO">
-        INSERT INTO consumption_amount(year, month, metropolitan_code, city_code, dong_code, dong_name, industry_code, industry_name, natinality, period, amount, consumption_amount_foreigner)
-        VALUES (#{year}, #{month}, #{metropolitan_code}, #{city_code}, #{dong_code}, #{dong_name}, #{industry_code}, #{industry_name}, #{nationality}, #{period}, #{count}, #{amount}, #{consumptionAmountForeigner})
+    <insert id="insert" useGeneratedKeys="true" keyProperty="id">
+        INSERT INTO consumption_amount_foreigner(year, month, metropolitan_code, metropolitan_name, city_code, city_name, dong_code, dong_name, industry_code, industry_name, nationality, period, count, amount)
+        VALUES (#{year}, #{month}, #{metropolitan_code}, #{metropolitan_name}, #{city_code}, #{city_name}, #{dong_code}, #{dong_name}, #{industry_code}, #{industry_name}, #{nationality}, #{period}, #{count}, #{amount})
     </insert>
 </mapper>
