+++ src/main/resources/sqlmapper/dailyFloatingPopulation.xml
@@ -4,13 +4,13 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="mapper.DFPMapper">
-    <select id="selectAll" resultType="DB.DTO.DailyFloatingPopulationDTO">
+    <select id="selectAll" resultType="persistence.DTO.DailyFloatingPopulationDTO">
         SELECT *
         FROM daily_floating_population
     </select>
 
-    <insert id="insert" useGeneratedKeys="true" keyProperty="id" parameterType="DB.DTO.DailyFloatingPopulationDTO">
-        INSERT INTO consumption_amount(year, month, dong_code, dong_name, time, daily_floating_population)
+    <insert id="insert" useGeneratedKeys="true" keyProperty="id" parameterType="persistence.DTO.DailyFloatingPopulationDTO">
+        INSERT INTO daily_floating_population(year, month, dong_code, dong_name, time, daily_floating_population)
         VALUES (#{year}, #{month}, #{dong_code}, #{dong_name}, #{time}, #{daily_floating_population})
     </insert>
 </mapper>
